Fix dot completion not working in middle of line
Today I learned something new about Python.
If the upper bound of a slice exceeds the length of the string, it's
automatically truncated instead of raising an error.

It turns out that I was slicing a string 1 character later
than needed and not realizing it, because it works perfectly if the
cursor is at the end of the string.